restless/requirements.txt
ansible==8.2.0
ansible-core==2.15.2
argcomplete==1.8.1
asgiref==3.6.0
attrs==20.3.0
autopep8==2.0.1
bcc==0.18.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
capstone==5.0.0rc2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
check50==3.3.7
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
cryptography==3.4.8
cs50==9.2.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-unused-media==0.2.2
djangorestframework==3.14.0
dnspython==2.1.0
docopt==0.6.2
evdev==1.4.0
exceptiongroup==1.1.2
fail2ban==0.11.2
filelock==3.6.0
Flask==2.2.2
gpg==1.16.0
greenlet==2.0.2
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
jeepney==0.7.1
jellyfish==0.9.0
Jinja2==3.1.2
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lib50==3.0.5
libvirt-python==8.0.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
more-itertools==8.10.0
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
oauthlib==3.2.0
olefile==0.46
openCVE @ file:///home/dgclasher/Development/projects/github/openCVE/dist/openCVE-0.1-py3-none-any.whl
packaging==23.0
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.2.0
plumbum==1.8.1
pop-transition==1.1.2
protobuf==3.12.4
protonvpn-cli==2.2.11
psutil==5.9.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pwntools==4.9.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
pyelftools==0.29
Pygments==2.14.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
pytest==7.4.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.7
python-gnupg==0.4.8
python-xlib==0.29
pythondialog==3.5.3
pytz==2022.7.1
pyxdg==0.27
PyYAML==6.0
repolib==2.2.1
repoman==1.4.0
reportlab==3.6.8
requests==2.28.2
requests-toolbelt==0.9.1
resolvelib==1.0.1
ROPGadget==7.2
rpyc==5.3.0
scapy==2.5.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
submit50==3.1.1
system76driver==20.4.79
systemd-python==234
termcolor==1.1.0
tomli==2.0.1
tqdm==4.64.1
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unicorn==2.0.1.post1
urllib3==1.26.14
vboxapi==1.0
virtualenv==20.16.7
wadllib==1.3.6
Werkzeug==2.2.2
wget==3.2
whitenoise==6.3.0
wifite==2.6.0
wikipedia==1.4.0
xdg==5
xkit==0.0.0
zipp==1.0.0
